pyproject.toml
[project]
name = "aqi-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "numpy>=2.4.4",
    "opencv-python>=4.9.0",
    "pillow>=12.2.0",
    "python-multipart>=0.0.26",
    "scikit-image>=0.22.0",
    "tensorflow>=2.16.1",
    "uvicorn>=0.44.0",
]
